requirements.txt
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.0.1
python-telegram-bot==20.0
APScheduler==3.11.0
influxdb==5.3.2
influxdb-client==1.46.0